ecosnap_backend/requirements.txt
Django>=4.2
djangorestframework
gunicorn
Pillow
celery
redis
psycopg2-binary
djangorestframework-simplejwt
graphene-django
graphene_file_upload
torch
torchvision
scikit-learn
pandas
numpy
matplotlib
transformers
django-celery-beat
pyfcm
